fsdfdsfs

+++ sprint4/problems/leave_game/solution/src/database/database_manager.h
@@ -6,14 +6,14 @@ namespace database {
     struct Record {
         std::string name;
         int score;
-        float play_time;
+        double play_time;
     };
 
     class DatabaseManager {
     public:
         DatabaseManager(const std::string& connection_info, std::size_t num_threads);
 
-        void SaveScore(const std::string& name, int score, float play_time);
+        void SaveScore(const std::string& name, int score, double play_time);
 
         std::vector<Record> GetRecords(int start, int item_limit);
